convert input to int32

+++ fade.py
@@ -10,7 +10,7 @@
 import sys
 from scipy.io.wavfile import read as readwav
 from scipy.io.wavfile import write as writewav
-from numpy import int16, power, log10, sin, linspace, pi
+from numpy import int16, int32, power, log10, sin, linspace, pi
 from os import mkdir
 import string
 
@@ -37,7 +37,10 @@ def main():
 		sys.exit(3)
 
 	print(f'opened {sys.argv[1]}, sample rate {input_sample_rate}')
-
+	
+	# convert input audio to int32
+	input_audio = input_audio.astype(int32)
+	
 	sample_count = len(input_audio)
 
 	# create fading profile as a sine wave from 0 dB to -(fade magnitude) dB
